chore(tests): moves search tests to tests.spec, so they will run one after another;
